+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- Target class in field decorator instead of prototype
+
 ## [1.0.0] 2018-04-21
 
 - Initial version